Remove relative import paths
The mix of relative paths seems to change
the Config include order which breaks testing
because the config global variables are reinitialized.

The non-relative imports are imho easier to read
anyway, and they give opportunity to make the rest
of the code more readable by removing the "as XXX"
renames in a followup commit.